+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ^^^^^^^^^^^^^^^^^^
 
 * Fix primitive type with metadata handling.
+* Fix multi-dimensional array handling.
 
 0.3.3 (2022-01-26)
 ^^^^^^^^^^^^^^^^^^